feat(service-betterportal): add capability support
Add support for adding new capabilities to the BetterPortal service. The
`addCapability` method was added to the `BetterPortalCallable` interface and
implemented in the `Service` class. The `BetterPortalCapabilityConfigurable`
enum was added to the `index.ts` file to define the configurable capabilities.
The `BetterPortalCapabilityInternal` enum was also added to define the internal
capabilities. The `BetterPortalCapability` type was added to define the
capabilities that can be used. The `capabilities` array was added to the
`Service` class to store the added capabilities. The `/bp/capabilities/` and
`/bp/capabilities/:capability/:optionalparam?/` routes were added to the
`Service` class to handle the capabilities.